Rename service, make this more explicitly a wrapper class
